Update tests - split exposure manager and FM acceptance tests
